refactored fixtures fro test_provider to fixtures.py
